requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
asyncpg==0.29.0
click==8.1.7
colorama==0.4.6
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
fastapi==0.109.2
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
ipykernel==6.29.2
ipython==8.21.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
matplotlib-inline==0.1.6
mysql==0.0.3
mysqlclient==2.2.3
nest-asyncio==1.6.0
packaging==23.2
parso==0.8.3
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
stack-data==0.6.3
starlette==0.36.3
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
